requirements.txt
python-dotenv==1.0.0
openai==0.27.8
facebook-scraper==0.2.59
python-telegram-bot==20.3
pre-commit==3.3.3
